Create shared package
It is practical to have a shared package in Val.

In addition the decision to create a new package called shared, was driven by the fact that both stegaEncode, used by useVal, in the react package and the ui package depended on parseRichTextSource which resulted in issues since parseRichTextSource was the UI package and the UI package must be "use client" only to avoid hitting errors since down-the-line it depended on Lexical which uses a createContext.

There might be other parts of the code that can fit this package, but for now, we are focusing on getting everything up and running.