Bare POC of Val UI from server
If we can do this, we can make all dependencies in Val UI devDependencies + less chances of things breaking. Also bundle sizes should be smaller (about 1mb compressed!).
In addition we can use the Val UI in outside of React all together!
These are advantages we BELIEVE this has currently. We might be wrong in some or all of them.

All in all, very promising stuff....