Move route to (val) and use a page to render the val app (fullscreen) - CURRENTLY WIP
Current commits breaks it. Needs to fixed after bundling. Also fails with the @react-refresh which needs a shim.

Having a (val) feels a lot cleaner than having it served on /api/val/static/edit which felt like it never could be the actual URL we should be using.

In addition this paves the way for us to do the same for the Overlay. In turn this means we can remove the runtime dependencies on Val UI modules which means we can move all UI deps to devDeps. That way we are less tied to the specific versions, in addition to having to download less data.
Also we believe we can make loading a lot faster, since the Val UI code will be prebuilt and easily cacheable (in the browser).