Pre-launch (last!) refactor: val.content => c.define val.file => c.file and c.rt.image (for richtext) val.richtext => c.richtext val.link => c.rt.link
The reason is that we want to possibly put more utils inside val and most of them would NOT be possible to use in content files. This way we split everything that was used in content files in c. and everything else goes into val.
We can also lint and make sure val. is not used inside content files

TODOs: patches, init script, tests, get example running, tag rt specific things,