WIP: Started on new lexer for RichText.
We need to parse RichText in @valbuild/core, which means it cannot be a dependency. Also: we do not need to support ALL markdown features: images, links, code blocks, escaped text is / will be interpolated so there is no need for it. Why not just support 100% markdown? First: there is no 100% markdown, second: we need to interpolate images (which is probably the most common use case) so what do we chose only do differently for val images or teach devs to use c.rt instead (we chose to teach devs to use c.rt :)?