Set default values to empty of...
Reason: Val should not create data where types are not correct. For certain schema types (image, file) this means that it would never validate, but it will not fail runtime unless one is expecting a value.
For keyOf it will currently possibly fail so we need to look into this. Feels like there's nothing we can do in this case?