Remove react/internal ValProvider, ValUI, ...
... Replaced by Script block.
Before moving to script block the host / client application would build the Val Overlay. This resulted in a lot of dependencies: a lot of packages (ca 320mb!) most coming from UI + required that the host app have the same versions of UI packages.

Some day Val might have a UI builder package that lets the host app build their application in React, but that is out of scope for now.