Introducing the val.modules file...
Users must now add modules that are editable to this file. This is the disadvantage.

The advantage is that Val no longer needs to evaluate files using QuickJS. So no need for QuickJS, no need for heavy caching in the content server, nor will the proxy need to load files from github (before patching). Now the proxy simply reads the val.modules files instead of loading this from github.
Overall there's less magic.
Everything will load much faster both in dev and prod. We also believe this is a concept that should be relatively familiar to many people, in particular those used to sanity. The val.modules files can be extended also be used to hide or give permissions on modules or how files are shown in the UI...

The plan is to test this out first before changing too much making it harder to revert (yes: we could git revert but that might not be easy once we start piling on things). That means we will keep QuickJS and the Service class that can evaluate files for now.

There's a lot of questions around whether this is a good idea.
Will it even work in proxy / prod mode? Can we get hot module replacement working when Val is enabled? Did it work before? What do we do now with the VS code extension - will it use quickjs or will it use something else? Finally, although we believe this makes Val less magical, it is another thing to remember - to add a editable module you must declare it val.modules - it will be something easy to forget. Some tooling is expected to be necessary to make up for it that, but are will this tooling hurt devex, will they be expensive to make?