Refactoring from Proxy- and LocalValServer to ValOps, ValServer2
Currently it seems to be working but no validation is done. Also no errors are handled.
The plan is to continue on this path and once ValOpsHttp is working and we can add changes in production,
we'll circle back and take care of this. The reason: it is urgent to get changes in prod up and running (blank needs this)