WIP: seems like we are almost there with useValState...
... Needs more testing so might be wrong.
Also there is quite a lot of jitter if the field is a controlled component. We might not need that, but we will do a bit more research and see if we cannot fix it so that even a input field (for example string), can have "controlled" values from sources, and emit patches.