Started on integrating the new layout with the new ValProvider...
There is a bunch of changes expect how state and UX looks like. One of the things is that fields get source path and are responsible on their own for computing their schema and getting their source.

The idea is to decouple the fields from their parents. I am worried about the perf, but for now it looks good