Fix PatchSets and use schema of types that do not require source to be analyzed...
The disadvantage of this approach is that we do not know the exact schema type at the patch path. The advantage is that we do not need the actual source at the current patch to be able to get the schemas.
This means that schemas with union will have potentially bigger patchsets (higher up towards module level)