Tweak amount of bits / hex chars used in file hash and validation hash...
Increased file hash from 10, very low risk of collision, to 12, extremely low risk. But reduced length of validation hash from 8 to 6 since the chance and consequences of a collision is very low.
Consequence of wrong file is pretty bad, but the validation basis is just that it might no longer be valid (which itself is an unlikely scenario)