FileMetadata is now very specific: must be a mimeType and that's it.
The idea earlier was that FileMetadata was the generic container and that all the other forms of metadata was subtypes of it.
Now the thinking is that: we have a specific FileMetadata and it is has mimeType (or whatever else). ImageMetadata has now "no relation" (though it didn't really have that earlier either).

What is the concrete benefits? The advantage is that users can auto-complete mimeType and / or alt, hotspot, ... in image / file