Fix issues of moving elements in a s.array(s.string())...
So we use useSourceAtPath instead of useShallowSourceAtPath.
Honestly we are not sure why this works better.
Even after fixing dep array issue in  useShallowSourceAtPath which we were sure would fix it, we still had issues...

Figured we might as well use useSourceAtPath since semantically that is more correct (we figure)