Update app JS and CSS urls to use versions directly instead of redirects...
... The main intention here is to make the UI load faster.
Earlier, we would use those which would redirect to a versioned path. This proved to be pretty slow (anecdotally: 150ms for the redirect then 50 ms to get the file). When using the actual versioned path, we should be able to skip first redirect probably saving 100ms at least. And that would be only for the first request, because every request that goes directly to the actual, versioned,  file, can be safely cached which is different than what we have before where we ALWAYS had to hit the redirect first and only then know that we could use the cached file.