Upload directly from client: avoid too large payloads, improve UX, stability...
... Currently patches with  images and files fails on Vercel  if they are above 4.5mb.
Many files and images are much bigger than that.

Therefore, we no longer upload patches with files to Vercel, rather we upload them to Val directly from the client.

This is also makes it possible to separate the uploading of files and creation of the patches, and makes it possible for us to add progress indicators and better error handling.

When using FS we do not have this problem (payloads being too large), but since we want to behave as similarly as possible we have created a new endpoint that emulates the HTTP behavior.