Peg version of core and shared...
We're desperately searching for a weird issue related to parsing Patch (using zod) in the UI. The patches fails to validate, although they look correct. Maybe (though it is not likely) it is because we are bundling the wrong version of shared? Hard to verify if that is the actual problem, since we are building then bundling the whole shebang...