Rename preview -> render...
... The method name of preview works best if you only change the actual preview of the list. However, we make it possible to change how a list is rendered in its totality.
Consider richtext. We want to change from s.richtext() being an array to just a html node. Then consider that we want to s.array(s.richtext()) as the new multiline default for richtext. Then we implicitly change the way the array is rendered, we do not preview the list items at all... Words matter so, that's why we change it. We have full control over our current users so it is relatively easy to fix.